added functionality to collect 3rd party requests #58
Added filter/logic so that requests made to firestore are not wriiten to the databse and so that requests made by and to the same host are not written to the databse. Added index in order to query for the site on which the third party request was made (i.e. the most recent Browser History entry with the same tabId and hostname as the request). Added server timestamp data type to Browser History for the aforementioned query. Added onSendHeaders listener in background.js to call function which adds the third party requests.